fixed relationships + added crud

+++ SpringProjects/TP-Spring/src/main/java/tn/esprit/tp1/exceptions/ContratNotFoundException.java
@@ -2,6 +2,7 @@ package tn.esprit.tp1.exceptions;
 
 
 import java.io.Serializable;
+
 import static java.text.MessageFormat.format;
 
 
@@ -9,7 +10,7 @@ public class ContratNotFoundException extends RuntimeException implements Serial
 
     private static final long serialVersionUID = 1L;
 
-    public ContratNotFoundException(Long id) {
+    public ContratNotFoundException(Integer id) {
         super(format("Contrat with id {0} could not be found !", id));
     }
